+++ src/routes/router.tsx
@@ -1,32 +1,12 @@
-import { createBrowserRouter, Outlet, redirect } from "react-router-dom"
-import { Nav } from "../Nav"
-import { guardedRoutes } from "./guardedRoutes"
-import { paths } from "./paths"
-import { localStorageKey } from "../utils/ids"
+import { createBrowserRouter, createHashRouter } from "react-router-dom"
+import { rootRoutes } from "./rootRoutes"
 
-export const isAuthenticated = () =>
-  localStorage.getItem(localStorageKey.jwtToken) !== null
+console.log(
+  process.env.NODE_ENV === "production"
+    ? "using HashRouter"
+    : "using BrowserRouter",
+)
+const wrapper =
+  process.env.NODE_ENV === "production" ? createHashRouter : createBrowserRouter
 
-export const logout = async () => {
-  localStorage.removeItem(localStorageKey.jwtToken)
-  return redirect(paths.signin)
-}
-
-export const router = createBrowserRouter([
-  {
-    path: "/",
-    element: (
-      <>
-        <Nav />
-        <Outlet />
-      </>
-    ),
-    children: guardedRoutes,
-    loader: isAuthenticated,
-    action: logout,
-  },
-  {
-    path: "*",
-    loader: () => redirect(paths.signin),
-  },
-])
+export const router = wrapper(rootRoutes)
